feat(container): improved pnr fix pr
ref: #MANAGER-14863
