--- packages/manager/apps/container/src/container/nav-reshuffle/sidebar/ProjectSelector/ProjectSelector.constants.tsx
@@ -1,2 +1,0 @@
-export const SEE_ALL_PROJECTS_ITEM_ID = 'seeAll';
-export const NEW_PROJECT_ITEM_ID = 'new';
